Scrollback stitches past herdr's read cap, and says so when it cannot
herdr returns at most 1000 rows per read and takes no offset, so the node grows
a deeper ring by matching the overlap between successive reads and appending
only what is new. Proven live at 1553 rows, every one of which is above what a
single read can return, with all 1600 markers accounted for and colour intact.

When output outruns the poll the reads share no overlap, and the two stretches
of history are not adjacent. Splicing them would make from_top and total_rows
fiction, so the node drops what it held, advances the base so absolute indices
stay true, and sets capped. A width change restarts the ring for a different
reason — stored rows were wrapped at the old width — and logs which happened.

Preserving history across a gap would need a per-segment from_top or a gap
sentinel; both are wire changes and neither is in v1. Adaptive polling is the
mitigation, the discard is the honest floor beneath it.

The desk-resize poll is now documented as the sole mechanism rather than a
backstop, and shortened to 3s accordingly.

+++ crates/kampr-term/src/grid.rs
@@ -62,7 +62,9 @@ impl Default for Cell {
 
 #[derive(Debug, Clone, Serialize)]
 pub struct RowDiff {
-    pub row: u16,
+    /// Wide enough for an absolute scrollback index: a ring outgrows 16 bits long before it
+    /// outgrows memory.
+    pub row: u32,
     pub cells: Vec<Cell>,
 }
 
@@ -146,7 +148,7 @@ impl Grid {
         for r in 0..self.rows {
             if std::mem::replace(&mut self.dirty[r as usize], false) {
                 out.push(RowDiff {
-                    row: r,
+                    row: r as u32,
                     cells: self.row(r).to_vec(),
                 });
             }
